Create component folders (#113)
Refactor folders arranged by components
