+++ gui/components/_control_bars/__init__.py
@@ -1,0 +1,1 @@
+from .model import ControlBars
