+++ gui/components/_dimensions/model.py
@@ -1,11 +1,11 @@
-from ..util.misc import (compute_max_shape as _compute_max_shape,
+from ...util.misc import (compute_max_shape as _compute_max_shape,
                          guess_metadata)
 from numpy import clip, integer, ndarray, append, insert, delete
 from copy import copy
 
 from vispy.util.event import EmitterGroup, Event
 
-from .qt import QtDimensions
+from .view import QtDimensions
 
 
 class Dimensions:
