+++ gui/components/_layers_list/view/__init__.py
@@ -1,0 +1,1 @@
+from .main import QtLayersPanel
