+++ gui/components/_window/__init__.py
@@ -4,10 +4,7 @@ wrap.
 """
 from PyQt5.QtWidgets import QApplication as QtApplication
 
-from ._layer_panel import QtLayerPanel
-from ._control_bars import QtControlBars
-from ._dimensions import QtDimensions
-from ._viewer import QtViewer
+from .view import Window
 
 from vispy import app
 app.use_app('pyqt5')
