+++ gui/layers/_base_layer/__init__.py
@@ -1,0 +1,2 @@
+from .model import Layer
+from .view import QtLayer
