+++ gui/layers/_base_layer/view.py
