+++ gui/layers/_image_layer/model.py
@@ -1,22 +1,22 @@
 import numpy as np
 from numpy import clip, integer, ndarray, append, insert, delete, empty
 from copy import copy
-from .vendored import cm
+from ..vendored import cm
 
-from ._base_layer import Layer
-from .._vispy.scene.visuals import Image as ImageNode
+from .._base_layer import Layer
+from ..._vispy.scene.visuals import Image as ImageNode
 
-from ..util import is_multichannel
-from ..util.interpolation import (interpolation_names,
+from ...util import is_multichannel
+from ...util.interpolation import (interpolation_names,
                                   interpolation_index_to_name as _index_to_name,  # noqa
                                   interpolation_name_to_index as _name_to_index)  # noqa
 
 from vispy import color
-from .colormaps import matplotlib_colormaps
+from ..colormaps import matplotlib_colormaps
 
-from ._register import add_to_viewer
+from .._register import add_to_viewer
 
-from .qt import QtImageLayer
+from .view import QtImageLayer
