+++ gui/layers/_image_layer/view.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QLabel, QComboBox
-from ._base_layer import QtLayer
+from .._base_layer import QtLayer
 
 
 class QtImageLayer(QtLayer):
