+++ gui/layers/_markers_layer/__init__.py
@@ -1,0 +1,1 @@
+from .model import Markers
