--- gui/layers/qt/__init__.py
@@ -1,7 +1,0 @@
-"""
-Custom Qt widgets that serve as native objects that the public-facing elements
-wrap.
-"""
-
-from ._image_layer import QtImageLayer
-from ._markers_layer import QtMarkersLayer
